Install as data by default
Installing as a Python module proper was problematic, and was causing Zim
to ignore the plugin with a lot of installation types. You could not, for
example, install to /usr/local if Zim is installed to /usr. The like-named
zim.plugins modules effectively shadow each other.

The correct fix is to allow Zim, not Python, to manage plugins: Zim also
searches $XDG_DATA_DIRS and $XDG_DATA_HOME for 3rd-party plugins.
Unfortunately we cannot allow installs into $XDG_DATA_HOME due to the way
DBus service activation works.

Ref: http://bazaar.launchpad.net/~jaap.karssenberg/zim/pyzim/view/head:/HACKING/Writing_plugins.txt

Closes dsboger/zimsearch#2.

Also use distutils, because that is normal for a Python extension. It makes
sense even if zimsearch is just data files now, although the absolute path
requirement does need special handling.

Remove NEWS: was blank, and distutils doesn't require it.

Remove autotools machinery, and update .gitignore to deal with distutils
build clutter.